pyproject.toml
[tool.poetry]
name = "django_shop"
version = "0.1.0"
description = ""
authors = ["Daniil <sidorenkodk@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.6"
ipython = "^8.4.0"
psycopg2 = "^2.9.3"
postgres = "^4.0"
Pillow = "^9.2.0"
social-auth-app-django = "^5.0.0"
django-environ = "^0.9.0"
pytest = "^7.1.2"
python-json-logger = "^2.0.4"
redis = "^4.3.4"
celery = "^5.2.7"
gevent = "^21.12.0"
flower = "^1.2.0"
django-debug-toolbar = "^3.6.0"
django-debug-toolbar-template-profiler = "^2.1.0"
django-filter = "^22.1"
django-crispy-forms = "^1.14.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
